requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic-settings==2.7.0
requests==2.32.3
tldextract==5.1.3
reportlab==4.1.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.1.1
email-validator==2.1.1
